+++ src/main/ruby/truffleruby/core/range.rb
@@ -49,10 +49,6 @@ class Range
   end
   private :initialize
 
-  private def endless?
-    Primitive.nil? self.end
-  end
-
   def ==(other)
     return true if equal? other
 
@@ -104,7 +100,7 @@ class Range
 
   private def bsearch_float(&block)
     normalized_begin = self.begin.to_f
-    normalized_end = endless? ? Float::INFINITY : self.end.to_f
+    normalized_end = Truffle::RangeOperations.endless?(self) ? Float::INFINITY : self.end.to_f
     normalized_end = normalized_end.prev_float if self.exclude_end?
     min = normalized_begin
     max = normalized_end
@@ -344,23 +340,23 @@ class Range
   alias_method :member?, :include?
 
   def ===(value)
-    include?(value)
+    cover?(value)
   end
 
   def inspect
-    result = "#{self.begin.inspect}#{exclude_end? ? "..." : ".."}#{endless? ? "" : self.end.inspect}"
+    result = "#{self.begin.inspect}#{exclude_end? ? "..." : ".."}#{Truffle::RangeOperations.endless?(self) ? "" : self.end.inspect}"
     Primitive.infect(result, self)
   end
 
   def last(n=undefined)
-    raise RangeError, 'cannot get the last element of endless range' if endless?
+    raise RangeError, 'cannot get the last element of endless range' if Truffle::RangeOperations.endless?(self)
     return self.end if Primitive.undefined? n
 
     to_a.last(n)
   end
 
   def max
-    raise RangeError, 'cannot get the maximum of endless range' if endless?
+    raise RangeError, 'cannot get the maximum of endless range' if Truffle::RangeOperations.endless?(self)
     return super if block_given? || (exclude_end? && !self.end.kind_of?(Numeric))
     return nil if self.end < self.begin || (exclude_end? && self.end == self.begin)
     return self.end unless exclude_end?
@@ -378,7 +374,7 @@ class Range
 
   def min
     return super if block_given?
-    return nil if (!endless? && self.end < self.begin) || (exclude_end? && self.end == self.begin)
+    return nil if (!Truffle::RangeOperations.endless?(self) && self.end < self.begin) || (exclude_end? && self.end == self.begin)
 
     self.begin
   end
@@ -452,33 +448,21 @@ class Range
   end
 
   def to_s
-    result = "#{self.begin}#{exclude_end? ? "..." : ".."}#{endless? ? "" : self.end}"
+    result = "#{self.begin}#{exclude_end? ? "..." : ".."}#{self.end}"
     Primitive.infect(result, self)
   end
 
   def cover?(value)
-    # MRI uses <=> to compare, so must we.
-
-    beg_compare = (self.begin <=> value)
-    return false unless beg_compare
-
-    if Comparable.compare_int(beg_compare) <= 0
-      return true if endless?
-      end_compare = (value <=> self.end)
-
-      if exclude_end?
-        return true if Comparable.compare_int(end_compare) < 0
-      else
-        return true if Comparable.compare_int(end_compare) <= 0
-      end
+    if value.kind_of?(Range)
+      Truffle::RangeOperations.range_cover?(self, value)
+    else
+      Truffle::RangeOperations.cover?(self, value)
     end
-
-    false
   end
 
   def size
     return nil unless self.begin.kind_of?(Numeric)
-    return Float::INFINITY if endless?
+    return Float::INFINITY if Truffle::RangeOperations.endless?(self)
 
     delta = self.end - self.begin
     return 0 if delta < 0